+++ routes/console.php
@@ -35,6 +35,7 @@ Schedule::command(Commands\PublishScheduledPostsCommand::class)->everyMinute()->
 # Cleanup
 Schedule::command(Commands\CleanUpDatabaseCommand::class)->dailyAt('02:00');
 Schedule::command(Commands\CleanFfmpegTmpCommand::class)->hourly();
+Schedule::command(Commands\CleanExpiredZipsCommand::class)->hourly()->withoutOverlapping();
 
 # WebDAV ingest
 Schedule::command(Commands\IngestWebDavCommand::class)->everyFifteenMinutes();