+++ routes/web.php
@@ -64,13 +64,14 @@ Route::middleware(SetPublicLocale::class)->group(function (): void {
 
     Route::get('/dropbox/connect', [DropboxController::class, 'connect'])->name('dropbox.connect');
     Route::get('/dropbox/callback', [DropboxController::class, 'callback'])->name('dropbox.callback');
-    Route::post('/zips/channel/{channel}', [ZipController::class, 'startForChannel'])->name('zips.channel.start');
+    Route::get('/offers/{assignment}/download', [ZipController::class, 'video'])->middleware('signed')->name('offers.video.download');
+    Route::post('/zips/channel/{channel}', [ZipController::class, 'startForChannel'])->middleware('signed')->name('zips.channel.start');
     /**
      * @deprecated Use /zips/channel/{channel} instead
      */
-    Route::post('/zips/{batch}/{channel}', [ZipController::class, 'start'])->name('zips.start');
-    Route::get('/zips/{id}/progress', [ZipController::class, 'progress'])->name('zips.progress');
-    Route::get('/zips/{id}/download', [ZipController::class, 'download'])->name('zips.download');
+    Route::post('/zips/{batch}/{channel}', [ZipController::class, 'start'])->middleware('signed')->name('zips.start');
+    Route::get('/zips/{id}/progress', [ZipController::class, 'progress'])->middleware('signed')->name('zips.progress');
+    Route::get('/zips/{id}/download', [ZipController::class, 'download'])->middleware('signed')->name('zips.download');
 
     Route::get('/action-tokens/approve/{purpose}/{token}', [TokenApprovalController::class, 'update'])
         ->name('tokens.update');